requirements.txt
rouge-score
openai
langchain
langchain-community
langgraph
# tiktoken
langchain-openai
praw
transformers
torch
torchvision
torchaudio
flask
flask-cors
validators
newspaper3k
lxml-html-clean
supabase
google-cloud-firestore
firebase-admin